3DProgramming
This program is for my Programming for 3D assignment and will be updated
thought out.

So far in this program, models of the player and enemies have been
added, collectable class has been added, but need to be implemented.
Physics has been looked into and have classes add, will leave physics
for now.